MINIFICPP-1561 - Shuffle classes
