+++ libminifi/include/properties/PropertiesFile.h
@@ -20,7 +20,7 @@
 #include <string>
 #include <vector>
 
-#include "utils/EncryptionUtils.h"
+#include "utils/crypto/EncryptionUtils.h"
 #include "utils/OptionalUtils.h"
 
 namespace org {
